+++ src/core/infra/factories/controllers/list-all-files.ts
@@ -1,5 +1,5 @@
-import { Controller } from '../../../core/application/domain/controller'
-import { ListFilesController } from '../../../presentation/controllers/list-all-files-controller'
+import { ListFilesController } from '../../../../api/presentation/controllers/list-all-files-controller'
+import { Controller } from '../../../application/domain/controller'
 import { makeListFilesUseCase } from '../usecases/list-all-files'
 
 export const makeListFilesController = (): Controller => {
