+++ src/core/infra/factories/controllers/list-file-by-id.ts
@@ -1,5 +1,5 @@
-import { Controller } from '../../../core/application/domain/controller'
-import { ListFileByIdController } from '../../../presentation/controllers/list-file-by-id-controller'
+import { ListFileByIdController } from '../../../../api/presentation/controllers/list-file-by-id-controller'
+import { Controller } from '../../../application/domain/controller'
 import { makeListFileUseCase } from '../usecases/list-file-by-id'
 
 export const makeListFileController = (): Controller => {
