+++ src/core/infra/factories/controllers/upload-files.ts
@@ -1,5 +1,5 @@
-import { Controller } from '../../../core/application/domain/controller'
-import { UploadFilesController } from '../../../presentation/controllers/multi-upload-controller'
+import { UploadFilesController } from '../../../../api/presentation/controllers/multi-upload-controller'
+import { Controller } from '../../../application/domain/controller'
 import { makeUploadFilesUseCase } from '../usecases/upload-files'
 
 export const makeUploadFilesController = (): Controller => {
