+++ src/core/infra/factories/usecases/delete-file.ts
@@ -1,7 +1,7 @@
 import {
   DeleteFileUseCase,
   IDeleteFileUseCase,
-} from '../../../core/application/usecases/delete-file-use-case'
+} from '../../../application/usecases/delete-file-use-case'
 import { s3 } from '../../aws'
 
 export const makeDeleteFileUseCase = (): IDeleteFileUseCase => {
