+++ src/core/infra/factories/usecases/list-all-files.ts
@@ -1,11 +1,11 @@
 import {
   IListFilesUseCase,
   ListFilesUseCase,
-} from '../../../core/application/usecases/list-files-use-case'
+} from '../../../application/usecases/list-files-use-case'
 import { s3 } from '../../aws'
-import { loggerService } from '../../config/logging/winston'
+import { logging } from '../../config/logging/winston'
 import { RedisImplementation } from '../../redis/redis'
 
 export const makeListFilesUseCase = (): IListFilesUseCase => {
-  return new ListFilesUseCase(s3, loggerService, new RedisImplementation())
+  return new ListFilesUseCase(s3, logging, new RedisImplementation())
 }
