+++ src/core/infra/factories/usecases/upload-file.ts
@@ -1,10 +1,10 @@
 import {
   IUploadFileUseCase,
   UploadFileUseCase,
-} from '../../../core/application/usecases/upload-file-use-case'
+} from '../../../application/usecases/upload-file-use-case'
 import { s3 } from '../../aws'
-import { loggerService } from '../../config/logging/winston'
+import { logging } from '../../config/logging/winston'
 
 export const makeUploadFileUseCase = (): IUploadFileUseCase => {
-  return new UploadFileUseCase(s3, 'bucket-name', loggerService)
+  return new UploadFileUseCase(s3, 'bucket-name', logging)
 }
