+++ src/core/infra/factories/usecases/upload-files.ts
@@ -1,10 +1,10 @@
 import {
   IUploadFilesUseCase,
   UploadFilesUseCase,
-} from '../../../core/application/usecases/multi-upload-file-use-case'
+} from '../../../application/usecases/multi-upload-file-use-case'
 import { s3 } from '../../aws'
-import { loggerService } from '../../config/logging/winston'
+import { logging } from '../../config/logging/winston'
 
 export const makeUploadFilesUseCase = (): IUploadFilesUseCase => {
-  return new UploadFilesUseCase(s3, loggerService)
+  return new UploadFilesUseCase(s3, logging)
 }
